+++ src/vue/filters/formatNumber.js
@@ -1,5 +1,0 @@
-import { globalize } from './globalize'
-
-export function formatNumber (value) {
-  return globalize('formats.number', { value })
-}